pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdownkeeper"
version = "0.1.0"
description = "LLM-optimized markdown documentation management service"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "MarkdownKeeper Contributors" }]
license = { file = "LICENSE" }
dependencies = ["tomli>=2.0.1; python_version < '3.11'", "watchdog>=3.0.0"]

[project.scripts]
mdkeeper = "markdownkeeper.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


[project.optional-dependencies]
embeddings = ["sentence-transformers>=2.2"]
faiss = ["faiss-cpu>=1.7", "numpy>=1.24"]
